Adding function definition for bigwigCompare
To complement the changes to coverage calculation, I want to have the option to compare individually processed/normalized coverage files. deepTools function bigwigCompare allows this, so I've added that and modified the sample sheet to allow both individual sample coverage maps and comparison coverage maps to be stored for use in other processes. bamCompare can still be used in make_tracks_opts, too!